Optionally delete env file after reading it
This is necessary as some (all?) resources want a pristine and empty
directory where they can copy files to

+++ pkg/jindra/tools/crij/cmd/crij/main.go
@@ -160,9 +160,9 @@ func main() {
 				os.Exit(1)
 			}
 		}
-		if *deleteEnvFileAfterRead {
+		if !os.IsNotExist(err) && *deleteEnvFileAfterRead {
 			if err := os.Remove(*envFile); err != nil {
-				fmt.Fprintf(os.Stderr, "error deleting env file: %s -- continuing anyways\n")
+				fmt.Fprintf(os.Stderr, "error deleting env file: %s -- continuing anyways\n", err)
 			}
 		}
 		crij.SimpleEnvFileToEnv(string(content))
